requirements.txt
flask==3.0.*
flask-sqlalchemy==3.1.*
flask-wtf==1.2.*
python-dotenv==1.0.*
openai>=1.14
anthropic>=0.25
google-generativeai>=0.3
